first version done

+++ MT/MT/Services/mysqlINSERT.cs
@@ -125,7 +125,7 @@ namespace MT.Services
 
         }
 
-        public async Task<bool> addProductOrder(bool istemp, DateTime dateTime, int branchid,int qty, int productid)
+        public async Task<bool> addProductOrder(bool istemp, DateTime dateTime, int branchid,double qty, int productid)
         {
             bool result = false;
             refreshQueryString();
